# [libc++] Optimize ranges::rotate for vector<bool>::iterator