This PR optimizes the performance of `std::ranges::rotate` for `vector<bool>::iterator`. The optimization yields a performance improvement of up to **2096x**.

```
-----------------------------------------------------------------------------------
Benchmark                                         Before        After   Improvement
-----------------------------------------------------------------------------------
rng::rotate(vector<bool>) (by 1/4)/32            73.4 ns      25.9 ns         2.8x
rng::rotate(vector<bool>) (by 1/4)/50             107 ns      25.8 ns         4.1x
rng::rotate(vector<bool>) (by 1/4)/64             124 ns      25.5 ns         4.9x
rng::rotate(vector<bool>) (by 1/4)/128            227 ns      26.3 ns         8.6x
rng::rotate(vector<bool>) (by 1/4)/256            439 ns      18.5 ns          24x
rng::rotate(vector<bool>) (by 1/4)/512            866 ns      18.9 ns          46x
rng::rotate(vector<bool>) (by 1/4)/1024          1698 ns      19.1 ns          89x
rng::rotate(vector<bool>) (by 1/4)/2048          3908 ns      27.4 ns         143x
rng::rotate(vector<bool>) (by 1/4)/4096         14057 ns      26.6 ns         528x
rng::rotate(vector<bool>) (by 1/4)/8192         32199 ns      36.9 ns         873x
rng::rotate(vector<bool>) (by 1/4)/16384        66743 ns      54.9 ns        1216x
rng::rotate(vector<bool>) (by 1/4)/32768       133562 ns      93.4 ns        1430x
rng::rotate(vector<bool>) (by 1/4)/65536       305561 ns       166 ns        1841x
rng::rotate(vector<bool>) (by 1/4)/131072      621524 ns       323 ns        1924x
rng::rotate(vector<bool>) (by 1/4)/262144     1230311 ns       587 ns        2096x
rng::rotate(vector<bool>) (by 1/4)/524288     2475596 ns      1454 ns        1703x
rng::rotate(vector<bool>) (by 1/4)/1048576    5003340 ns      3177 ns        1575x
rng::rotate(vector<bool>) (by 51%)/32            15.0 ns      27.3 ns       -0.45x
rng::rotate(vector<bool>) (by 51%)/50            15.1 ns      27.6 ns       -0.45x
rng::rotate(vector<bool>) (by 51%)/64            15.2 ns      27.1 ns       -0.44x
rng::rotate(vector<bool>) (by 51%)/128            268 ns      29.2 ns           9x
rng::rotate(vector<bool>) (by 51%)/256            646 ns      34.5 ns          19x
rng::rotate(vector<bool>) (by 51%)/512           1604 ns      39.2 ns          41x
rng::rotate(vector<bool>) (by 51%)/1024          3397 ns      66.3 ns          51x
rng::rotate(vector<bool>) (by 51%)/2048          7965 ns      87.0 ns          92x
rng::rotate(vector<bool>) (by 51%)/4096         16475 ns       156 ns         106x
rng::rotate(vector<bool>) (by 51%)/8192         35793 ns       249 ns         144x
rng::rotate(vector<bool>) (by 51%)/16384        73247 ns       783 ns          94x
rng::rotate(vector<bool>) (by 51%)/32768       150151 ns      1223 ns         123x
rng::rotate(vector<bool>) (by 51%)/65536       296511 ns      2197 ns         135x
rng::rotate(vector<bool>) (by 51%)/131072      596007 ns      3951 ns         151x
rng::rotate(vector<bool>) (by 51%)/262144     1241285 ns      7553 ns         164x
rng::rotate(vector<bool>) (by 51%)/524288     2660962 ns     14659 ns         182x
rng::rotate(vector<bool>) (by 51%)/1048576    5295160 ns     28842 ns         184x
```

Closes #64038.

(Note: I've adopted the same benchmark code `modifying/rotate.bench.cpp` from #127354 for consistency. This ensures minimum conflicts when merging them.)